+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @swapkit/tokens
 
+## 4.1.1
+
+### Patch Changes
+
+- [`82d0fd8`](https://github.com/swapkit/SwapKit/commit/82d0fd8f1072d7341f45e3d07114181cec7500e1) Thanks [@ice-chillios](https://github.com/ice-chillios)! - Update tokens
+
+## 4.1.0
+
+### Minor Changes
+
+- [`2f181c6`](https://github.com/swapkit/SwapKit/commit/2f181c6d891d963854c5b7702a3de583c38d9317) Thanks [@ice-chillios](https://github.com/ice-chillios)! - Updates to isolated modules & wallet extensions extraction
+
 ## 4.0.43
 
 ### Patch Changes